Added new directories for project layers and renamed projects
