+++ .gitignore
@@ -2,3 +2,4 @@ bin/
 obj/
 .idea/
 PasswordToDb.txt
+.vs/
