+++ AnimeNotesAPI/Notebook.AnimeNotesAPI.csproj.Backup.tmp
@@ -1,0 +1,17 @@
+<Project Sdk="Microsoft.NET.Sdk.Web">
+
+    <PropertyGroup>
+        <TargetFramework>net9.0</TargetFramework>
+        <Nullable>enable</Nullable>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <RootNamespace>Notebook.AnimeNotesAPI</RootNamespace>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\BusinessLayer\BusinessLayer.csproj" />
+      <ProjectReference Include="..\Domain\Domain.csproj" />
+      <ProjectReference Include="..\Notebook.BusinessLayer\BusinessLayer.csproj" />
+      <ProjectReference Include="..\Notebook.Domain\Domain.csproj" />
+    </ItemGroup>
+
+</Project>
