+++ AnimeNotesAPI/Notebook.AnimeNotesAPI.csproj.user
