+++ AnimeNotesAPI/Properties/launchSettings.json
