+++ AnimeNotesAPI/appsettings.Development.json
