+++ AnimeNotesAPI/appsettings.json
