+++ BusinessLayer/Interfaces/ILog.cs
@@ -1,0 +1,6 @@
+namespace BusinessLayer.Interfaces;
+
+public interface ILog
+{
+    public void Log(string message);
+}
