+++ BusinessLayer/Interfaces/INoteValidator.cs
@@ -1,0 +1,8 @@
+using Domain.Models;
+
+namespace BusinessLayer.Interfaces;
+
+public interface INoteValidator
+{
+    bool Validate(Note note);
+}
