+++ Domain/Interfaces/INoteProvider.cs
@@ -1,0 +1,11 @@
+using Domain.Models;
+
+namespace Domain.Interfaces;
+
+public interface INoteProvider
+{
+    Note Create();
+    Note GetById(Guid id);
+    bool Update(Guid id, Note updatedNote);
+    public bool Delete(Guid id);
+}
