+++ Domain/Models/Enums/WatchingStatus.cs
@@ -1,0 +1,8 @@
+namespace Domain.Models.Enums;
+
+public enum WatchingStatus
+{
+    Wished,
+    InProcess,
+    Watched
+}
