--- Notebook.AnimeNotesAPI/Logger/ConsoleLog.cs
@@ -1,9 +1,0 @@
-namespace Notebook.AnimeNotesAPI;
-
-public class ConsoleLog : ILogger
-{
-    public void Log(string message)
-    {
-        Console.WriteLine(message);
-    }
-}
