--- Notebook.AnimeNotesAPI/Notebook.AnimeNotesAPI.csproj
@@ -1,10 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-    <PropertyGroup>
-        <TargetFramework>net9.0</TargetFramework>
-        <Nullable>enable</Nullable>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <RootNamespace>Notebook.AnimeNotesAPI</RootNamespace>
-    </PropertyGroup>
-
-</Project>
