--- Notebook.AnimeNotesAPI/Program.cs
@@ -1,6 +1,0 @@
-var builder = WebApplication.CreateBuilder(args);
-var app = builder.Build();
-
-app.MapGet("/", () => "Hello World!");
-
-app.Run();
