--- Notebook.BusinessLayer/Interfaces/ILogger.cs
@@ -1,6 +1,0 @@
-namespace Notebook.BusinessLayer.Interfaces;
-
-public interface ILogger
-{
-    public void Log(string message);
-}
