--- Notebook.BusinessLayer/Interfaces/INoteProvider.cs
@@ -1,11 +1,0 @@
-using Notebook.BusinessLayer.DomainModels;
-
-namespace Notebook.BusinessLayer.Interfaces;
-
-public interface INoteProvider
-{
-    Note Create();
-    Note GetById(Guid id);
-    bool Update(Guid id, Note updatedNote);
-    bool Delete(Guid id);
-}
