--- Notebook.BusinessLayer/Interfaces/INoteValidator.cs
@@ -1,8 +1,0 @@
-using Notebook.BusinessLayer.DomainModels;
-
-namespace Notebook.BusinessLayer.Interfaces;
-
-public interface INoteValidator
-{
-    bool Validate(Note note);
-}
