--- Notebook.Core/Models/Enums/Genres.cs
@@ -1,10 +1,0 @@
-namespace Notebook.BusinessLayer.Enums;
-
-public enum Genres
-{
-    Horror, Comedy, Thriller, Detective, Action, Adventure, Drama,
-    Fantasy, SciFi, Mystery, Supernatural, Psychological, Romance,
-    SliceOfLife, Mecha, Sports, Music, Historical, War, Isekai,
-    Ecchi, Harem, ReverseHarem, Parody, Games, School, Work,
-    Cooking, Beastmen
-}
