--- Notebook.Core/Models/Enums/WatchingStatus.cs
@@ -1,8 +1,0 @@
-namespace Notebook.BusinessLayer.Enums;
-
-public enum WatchingStatus
-{
-    Wished,
-    InProcess,
-    Watched
-}
