--- Notebook.Infrastructure/DAL/Classes/NoteProvider.cs
@@ -1,28 +1,0 @@
-using Notebook.BusinessLayer.DomainModels;
-using Notebook.BusinessLayer.Interfaces;
-
-namespace Notebook.Infrastructure.DAL.Classes;
-
-//TODO Implement Class
-public class NoteProvider : INoteProvider
-{
-    public Note Create()
-    {
-        throw new NotImplementedException();
-    }
-
-    public Note GetById(Guid id)
-    {
-        throw new NotImplementedException();
-    }
-
-    public bool Update(Guid id, Note updatedNote)
-    {
-        throw new NotImplementedException();
-    }
-
-    public bool Delete(Guid id)
-    {
-        throw new NotImplementedException();
-    }
-}
