--- Notebook.Infrastructure/DAL/Interfaces/INoteProvider.cs
@@ -1,6 +1,0 @@
-namespace Notebook.Infrastructure.DAL.Interfaces;
-
-public interface INoteProvider
-{
-    
-}
